+++ config/webpack.config.dev.js
@@ -47,7 +47,7 @@ module.exports = {
         }
     },
     entry: {
-        base: './src/assets/js/common/base.js',
+        base: './src/assets/js/common/base.ts',
         ...getEntry('./src/assets/js/view'),
         ...getEntry('./src/assets/js/view', '.ts')
     },
