+++ src/assets/js/common/base.ts
